requirements.txt
streamlit
streamlit-autorefresh
azure-ai-vision-imageanalysis
streamlit-js-eval
streamlit-js-eval
